refactor(core): split session library files
split session library files
