+++ packages/core/src/libraries/session/interaction.ts
@@ -1,0 +1,39 @@
+import type { Context } from 'koa';
+import type { InteractionResults, Provider } from 'oidc-provider';
+
+export const updateInteractionResult = async (
+  ctx: Context,
+  provider: Provider,
+  result: InteractionResults,
+  merge = false
+) => {
+  // The "mergeWithLastSubmission" will only merge current request's interaction results,
+  // which is stored in ctx.oidc, we need to merge interaction results in two requests,
+  // have to do it manually
+  // refer to: https://github.com/panva/node-oidc-provider/blob/c243bf6b6663c41ff3e75c09b95fb978eba87381/lib/actions/authorization/interactions.js#L106
+  const details = merge ? await provider.interactionDetails(ctx.req, ctx.res) : undefined;
+
+  return provider.interactionResult(
+    ctx.req,
+    ctx.res,
+    {
+      // Merge with current result
+      ...details?.result,
+      ...result,
+    },
+    {
+      mergeWithLastSubmission: merge,
+    }
+  );
+};
+
+export const assignInteractionResults = async (
+  ctx: Context,
+  provider: Provider,
+  result: InteractionResults,
+  merge = false
+) => {
+  const redirectTo = await updateInteractionResult(ctx, provider, result, merge);
+
+  ctx.body = { redirectTo };
+};
