+++ packages/core/src/routes/interaction/actions/submit-interaction.test.ts
@@ -29,7 +29,7 @@ const getLogtoConnectorById = jest
   .fn()
   .mockResolvedValue({ metadata: { target: 'logto' }, dbEntry: { syncProfile: true } });
 
-const { assignInteractionResults } = mockEsm('#src/libraries/session.js', () => ({
+const { assignInteractionResults } = mockEsm('#src/libraries/session/index.js', () => ({
   assignInteractionResults: jest.fn(),
 }));
